Start moving struct tests
