+++ python/cudf/cudf/tests/series/indexing/test_getitem.py
@@ -1,0 +1,66 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize(
+    "data, idx, expected",
+    [
+        (
+            [{"f2": {"a": "sf21"}, "f1": "a"}, {"f1": "sf12", "f2": None}],
+            0,
+            {"f1": "a", "f2": {"a": "sf21"}},
+        ),
+        (
+            [
+                {"f2": {"a": "sf21"}},
+                {"f1": "sf12", "f2": None},
+            ],
+            0,
+            {"f1": cudf.NA, "f2": {"a": "sf21"}},
+        ),
+        (
+            [{"a": "123"}, {"a": "sf12", "b": {"a": {"b": "c"}}}],
+            1,
+            {"a": "sf12", "b": {"a": {"b": "c"}}},
+        ),
+    ],
+)
+def test_nested_struct_extract_host_scalars(data, idx, expected):
+    series = cudf.Series(data)
+
+    def _nested_na_replace(struct_scalar):
+        """
+        Replace `cudf.NA` with `None` in the dict
+        """
+        for key, value in struct_scalar.items():
+            if value is cudf.NA:
+                struct_scalar[key] = None
+        return struct_scalar
+
+    assert _nested_na_replace(series[idx]) == _nested_na_replace(expected)
+
+
+def test_nested_struct_from_pandas_empty():
+    # tests constructing nested structs columns that would result in
+    # libcudf EMPTY type child columns inheriting their parent's null
+    # mask. See GH PR: #10761
+    pdf = pd.Series([[{"c": {"x": None}}], [{"c": None}]])
+    gdf = cudf.from_pandas(pdf)
+
+    assert_eq(pdf, gdf)
+
+
+def test_struct_int_values():
+    series = cudf.Series(
+        [{"a": 1, "b": 2}, {"a": 10, "b": None}, {"a": 5, "b": 6}]
+    )
+    actual_series = series.to_pandas()
+
+    assert isinstance(actual_series[0]["b"], int)
+    assert isinstance(actual_series[1]["b"], type(None))
+    assert isinstance(actual_series[2]["b"], int)
