--- python/cudf/cudf/tests/structs/test_struct_methods.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
